+++ components/service/glean/src/main/java/mozilla/components/service/glean/storages/StringSetsStorageEngine.kt
@@ -14,30 +14,30 @@ import mozilla.components.support.ktx.android.org.json.toList
 import org.json.JSONArray
 
 /**
- * This singleton handles the in-memory storage logic for string lists. It is meant to be used by
- * the Specific String List API and the ping assembling objects.
+ * This singleton handles the in-memory storage logic for string sets. It is meant to be used by
+ * the Specific String Set API and the ping assembling objects.
  *
  * This class contains a reference to the Android application Context. While the IDE warns
  * us that this could leak, the application context lives as long as the application and this
  * object. For this reason, we should be safe to suppress the IDE warning.
  */
 @SuppressLint("StaticFieldLeak")
-internal object StringListsStorageEngine : StringListsStorageEngineImplementation()
+internal object StringSetsStorageEngine : StringSetsStorageEngineImplementation()
 
-internal open class StringListsStorageEngineImplementation(
-    override val logger: Logger = Logger("glean/StringsListsStorageEngine")
-) : GenericStorageEngine<List<String>>() {
+internal open class StringSetsStorageEngineImplementation(
+    override val logger: Logger = Logger("glean/StringsSetsStorageEngine")
+) : GenericStorageEngine<Set<String>>() {
     companion object {
-        // Maximum length of any list
-        const val MAX_LIST_LENGTH_VALUE = 20
-        // Maximum length of any string in the list
+        // Maximum size of any set
+        const val MAX_SET_SIZE_VALUE = 20
+        // Maximum length of any string in the set
         const val MAX_STRING_LENGTH = 50
     }
 
-    override fun deserializeSingleMetric(metricName: String, value: Any?): List<String>? {
+    override fun deserializeSingleMetric(metricName: String, value: Any?): Set<String>? {
         /*
-        Since SharedPreferences doesn't directly support storing of List<> types, we must use
-        an intermediate JSONArray which can be deserialized and converted back to List<String>.
+        Since SharedPreferences doesn't directly support storing of Set<> types, we must use
+        an intermediate JSONArray which can be deserialized and converted back to Set<String>.
         Using JSONArray introduces a possible issue in that it's constructor will still properly
         convert a stringified JSONArray into an array of Strings regardless of whether the values
         have been properly quoted or not.  For example, [1,2,3] is as valid just like
@@ -48,7 +48,7 @@ internal open class StringListsStorageEngineImplementation(
         */
         return (value as? String)?.let {
             try {
-                return@let JSONArray(it).toList()
+                return@let JSONArray(it).toList<String>().toSet()
             } catch (e: org.json.JSONException) {
                 return@let null
             }
@@ -58,21 +58,21 @@ internal open class StringListsStorageEngineImplementation(
     override fun serializeSingleMetric(
         userPreferences: SharedPreferences.Editor?,
         storeName: String,
-        value: List<String>,
+        value: Set<String>,
         extraSerializationData: Any?
     ) {
-        // Since SharedPreferences doesn't directly support storing of List<> types, we must use
+        // Since SharedPreferences doesn't directly support storing of Set<> types, we must use
         // an intermediate JSONArray which can be serialized to a String type and stored.
         val jsonArray = JSONArray(value)
         userPreferences?.putString(storeName, jsonArray.toString())
     }
 
     /**
-     * Appends to an existing string list in the desired stores.  If the store or list doesn't exist
+     * Appends to an existing string set in the desired stores.  If the store or set doesn't exist
      * then it is created and added to the desired stores.
      *
      * @param metricData object with metric settings
-     * @param value the string list value to add
+     * @param value the string value to add
      */
     fun add(
         metricData: CommonMetricData,
@@ -91,36 +91,40 @@ internal open class StringListsStorageEngineImplementation(
             it
         }
 
-        // Use a custom combiner to add the string to the existing list rather than overwriting
-        super.recordMetric(metricData, listOf(truncatedValue), null) { currentValue, newValue ->
+        // Use a custom combiner to add the string to the existing set rather than overwriting
+        super.recordMetric(metricData, setOf(truncatedValue), null) { currentValue, newValue ->
             currentValue?.let {
-                if (it.count() + 1 > MAX_LIST_LENGTH_VALUE) {
+                // Report an error when the set is too big.
+                if (it.count() + 1 > MAX_SET_SIZE_VALUE) {
                     recordError(
                         metricData,
                         ErrorType.InvalidValue,
-                        "String list length of ${it.count() + 1} exceeds maximum of $MAX_LIST_LENGTH_VALUE",
+                        "String set length of ${it.count() + 1} exceeds maximum of $MAX_SET_SIZE_VALUE",
                         logger
                     )
+                    // Return the unchanged set.
+                    return@let it
                 }
-
-                it + newValue.take(MAX_LIST_LENGTH_VALUE - it.count())
+                // Add the item to the set.
+                it + newValue
             } ?: newValue
         }
     }
 
     /**
-     * Sets a string list in the desired stores. This function will replace the existing list or
-     * create a new list if it doesn't already exist. To add or append to an existing list, use
+     * Sets a string set in the desired stores. This function will replace the existing set or
+     * create a new set if it doesn't already exist. To add or append to an existing set, use
      * [add] function.
      *
      * @param metricData object with metric settings
-     * @param value the string list value to record
+     * @param value the string set value to record
      */
+    @Suppress("LongMethod")
     fun set(
         metricData: CommonMetricData,
-        value: List<String>
+        value: Set<String>
     ) {
-        val stringList = value.map {
+        val stringSet = value.map {
             if (it.length > MAX_STRING_LENGTH) {
                 recordError(
                     metricData,
@@ -132,15 +136,15 @@ internal open class StringListsStorageEngineImplementation(
             it.take(MAX_STRING_LENGTH)
         }
 
-        if (stringList.count() > MAX_LIST_LENGTH_VALUE) {
+        if (stringSet.count() > MAX_SET_SIZE_VALUE) {
             recordError(
                 metricData,
                 ErrorType.InvalidValue,
-                "String list length of ${value.count()} exceeds maximum of $MAX_LIST_LENGTH_VALUE",
+                "String set length of ${value.count()} exceeds maximum of $MAX_SET_SIZE_VALUE",
                 logger
             )
         }
 
-        super.recordMetric(metricData, stringList.take(MAX_LIST_LENGTH_VALUE))
+        super.recordMetric(metricData, stringSet.take(MAX_SET_SIZE_VALUE).toSet())
     }
 }